feat: implement cli headers
